- NNUEのKP256のような、K + P型でうまく評価関数ファイルが読み込めなかったのを修正。

+++ source/eval/nnue/features/k.h
@@ -20,8 +20,10 @@ class K {
   static constexpr const char* kName = "K";
   // 評価関数ファイルに埋め込むハッシュ値
   static constexpr std::uint32_t kHashValue = 0xD3CEE169u;
+
   // 特徴量の次元数
-  static constexpr IndexType kDimensions = SQ_NB * 2;
+  static constexpr IndexType kDimensions = uint32_t(SQ_NB) * 2;
+
   // 特徴量のうち、同時に値が1となるインデックスの数の最大値
   static constexpr IndexType kMaxActiveDimensions = 2;
   // 差分計算の代わりに全計算を行うタイミング
